+++ src/main-handlers/api/build-table.js
@@ -1,0 +1,21 @@
+module.exports = ({ table }, params) => {
+  const headers = ['key', 'default', 'description']
+  return table([
+    headers.map(header => header.charAt(0).toUpperCase() + header.slice(1)),
+    ...params.map(buildLine(headers))
+  ],
+  { align: headers.map(header => header === 'description' ? 'l' : 'c') }
+  )
+}
+
+function buildLine (headers) {
+  return param => headers.map(header => {
+    if (!param[header]) {
+      return ''
+    }
+    if (header === 'key' || header === 'default') {
+      return `\`${param[header]}\``
+    }
+    return param[header]
+  })
+}
